+++ ChangeLog.md
@@ -22,7 +22,11 @@
 ## Goal - 0.2.0.0
 This version of the lexer and parser should be done entirely with the builtin libraries for parsing in Haskell. 
 
-### 0.1.1.0 -- 2016-12-27 
+#### 0.1.1.1  -- 2016-12-28
+* The reminder operator is added to the parser 
+* Error when parsing parentizies (Had mixed up the left and right)
+
+### 0.1.1.0   -- 2016-12-27 
 * Evaluator now has defined functions for evaluating programs. 
     * Evaluator can evauale basic arithmics
 * Basic tests for the evaluator has been created